Update endpoints.json, fix query bug to return int instead string for comment_count.

+++ app/Model/articleModel.js
@@ -11,7 +11,7 @@ exports.fetchArticleByID = async (article_id) => {
     articles.created_at, 
     articles.votes, 
     articles.article_img_url,
-    COUNT(comments.comment_id) AS comment_count
+    COUNT(comments.comment_id)::int AS comment_count
     FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id
     WHERE articles.article_id = $1
@@ -33,7 +33,7 @@ exports.fetchAllArticles = async (topic) => {
     articles.created_at, 
     articles.votes, 
     articles.article_img_url,
-    COUNT(comments.comment_id) AS comment_count
+    COUNT(comments.comment_id)::int AS comment_count
     FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id
   `
